+++ index.js
@@ -85,32 +85,32 @@ function _segseg (out, p1, p2, p3, p4) {
 }
 
 
-export default function segseg (out, p1, p2, p3, p4) {
+export default function segseg (out, p1, p2, p3, p4, epsilon=0) {
   const result = _segseg(out, p1, p2, p3, p4)
 
   if (result === DO_INTERSECT)
     return result
 
   // handle colinear cases and when a line segment endpoint lies on the other segment
-  if (segpoint(p1, p3, p4)) {
+  if (segpoint(p1, p3, p4, epsilon)) {
     out[0] = p1[0]
     out[1] = p1[1]
     return true
   }
 
-  if (segpoint(p2, p3, p4)) {
+  if (segpoint(p2, p3, p4, epsilon)) {
     out[0] = p2[0]
     out[1] = p2[1]
     return true
   }
 
-  if (segpoint(p3, p1, p2)) {
+  if (segpoint(p3, p1, p2, epsilon)) {
     out[0] = p3[0]
     out[1] = p3[1]
     return true
   }
 
-  if (segpoint(p4, p1, p2)) {
+  if (segpoint(p4, p1, p2, epsilon)) {
     out[0] = p4[0]
     out[1] = p4[1]
     return true